Timer cleanup continued
TIME_MILLIS_T reduced to 16 bits, MICROS to 32 bits
Seems to work fine with wraparounds now

+++ libraries/AP_Compass/examples/AlphaPilot/Time.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-typedef uint64_t VP_TIME_MICROS_T;
-typedef uint32_t VP_TIME_MILLIS_T;
+typedef uint32_t VP_TIME_MICROS_T;
+typedef uint16_t VP_TIME_MILLIS_T;
 
 typedef struct VPInertiaTimer {
   bool *state;
